Complete overhaul of menu
Created a new menu that allows you to scroll through list of songs, collapse groups, and displays the banner image of the selected song. Next I will have to make it display other info like title and BPM, play a snippit of the song, and add a play button to take you in game.