Music is functioning!
I have gotten it so when you select music on the main menu it will play a snippet of the song. This is a huge checkpoint for me so I am incredibly excited. I have also done much more code clean up and begun working on the actual note placement code.